asgn_4/pyproject.toml
[project]
name = "asgn-4"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pycryptodome>=3.23.0",
    "textdistance>=4.6.3",
]
